% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condenseGsea.R
\name{condense_gsea}
\alias{condense_gsea}
\title{add redundancy information to a GSEA table}
\usage{
condense_gsea(gsea, colname = "genes", sep = ",", similarity = 0.9)
}
\arguments{
\item{gsea}{data.frame, where each row is a gene set and there is at least one column with a string for each row. This string contains all genes for that set}

\item{colname}{string, naming the column in which the genes are listed}

\item{sep}{charachter. What genes in the column string? E.g. if a cell of the column would look like "HOXA9,HOXA3", you need to set sep=","}

\item{similarity}{number, what has to be the minimum gene overlap between two terms in order for one to be regarded as redundant?}
}
\value{
data.frame similar to the input, but with 3 added columns: condenseID has a simple numeric ID for each row. condenseChildren lists the IDs of all sets that were eaten by this set. condenseDropout states whether or not this set was eaten itself.
}
\description{
add redundancy information to a GSEA table
}
\examples{
library(gseaCondenser)

gsea <- gseaCondenser::myGsea
gsea <- condense_gsea(gsea, similarity=0.3)
head(gsea)
}
