% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqGsea.R
\name{freqGsea}
\alias{freqGsea}
\title{frequence table for GSEA terms}
\usage{
freqGsea(
  pathways,
  sep = "_",
  bias = NULL,
  clean = T,
  cleanII = T,
  removeIncluding = "",
  removeThis = ""
)
}
\arguments{
\item{pathways}{character vector. Upper or lower case makes no matter, since everything will be converted to uppercase anyway.}

\item{sep}{charachter. What separates words in a set name? E.g. if terms look like "KEGG_CELL_CYCLE", you need to set sep="_"}

\item{bias}{numeric vector of the same length as the pathways vector The pathways will be counted x times (x = bias). The numbers will be rounded to integers.}

\item{clean}{boolean, should common words and wordcombinations including these common words be removed? These common words include e.g. "the","of", etc. but also "GO","KEGG","REACTOME" and "PID"}

\item{cleanII}{boolean, should unspecific words (but not their combination with others) be removed? E.g. "cellular","activation","alpha" etc. Combinations like il_alpha will not be removed.}

\item{removeIncluding}{character vector. Listed words and combinations that include them will be removed.}

\item{removeThis}{character vector. Listed words will be removed, but not combinations that include them.}
}
\value{
Data.frame containing one column "word" and one column "freq", showing how many times one word (or word-combination) came up.
}
\description{
frequence table for GSEA terms
}
\examples{
library(gseaCondenser)
library(wordcloud2)

myfreq <- freqGsea(names(gseaCondenser::mySetlist))
wordcloud2(myfreq)
}
